+++ Magna/src/main/java/de/morigm/magna/api/language/Language.java
@@ -4,7 +4,7 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.Properties;
 
-import de.morigm.magna.Main;
+import de.morigm.magna.api.Magna;
 import de.morigm.magna.api.helper.LoadHelper;
 
 public class Language implements LoadHelper
@@ -25,7 +25,7 @@ public class Language implements LoadHelper
 		prop = new Properties();
 		try 
 		{
-			prop.load(new FileInputStream(Main.getInstance().getLanguageLoader().getLanguage()));
+			prop.load(new FileInputStream(Magna.getFolders().getLanguageFile()));
 		}
 		catch (IOException e) {e.printStackTrace();}
 	}