+++ Magna/src/main/java/de/morigm/magna/api/user/User.java
@@ -1,0 +1,6 @@
+package de.morigm.magna.api.user;
+
+public class User 
+{
+
+}