+++ Magna/src/main/java/de/morigm/magna/config/AutoEditConfig.java
@@ -5,7 +5,7 @@ import java.io.File;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 
-import de.morigm.magna.Main;
+import de.morigm.magna.api.Magna;
 import de.morigm.magna.api.helper.FileHelper;
 import de.morigm.magna.api.helper.LoadHelper;
 import lombok.Getter;
@@ -21,10 +21,10 @@ public class AutoEditConfig implements LoadHelper
 	@Override
 	public void load() 
 	{
-		this.file = new File(Main.getInstance().getDataFolder(), "autoedit.yml");
+		this.file = Magna.getFolders().getAutoEditFile();
 		FileHelper.createFileIfNotExists(file);
 		this.config = YamlConfiguration.loadConfiguration(file);
-		this.filePlayer = new File(Main.getInstance().getDataFolder(), "playerautoedit.yml");
+		this.filePlayer = Magna.getFolders().getPlayerAutoEditFile();
 		FileHelper.createFileIfNotExists(filePlayer);
 		this.configPlayer = YamlConfiguration.loadConfiguration(filePlayer);
 	}