+++ Magna/src/main/java/de/morigm/magna/config/GroupConfig.java
@@ -5,7 +5,7 @@ import java.io.File;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 
-import de.morigm.magna.Main;
+import de.morigm.magna.api.Magna;
 import de.morigm.magna.api.helper.ConfigHelper;
 import de.morigm.magna.api.helper.FileHelper;
 
@@ -18,7 +18,7 @@ public class GroupConfig implements ConfigHelper
 	@Override
 	public void load() 
 	{
-		this.file = new File(Main.getInstance().getDataFolder(),"groups.yml");
+		this.file = Magna.getFolders().getGroupsFile();
 		FileHelper.createFileIfNotExists(file);
 		this.config = YamlConfiguration.loadConfiguration(file);
 	}