+++ Magna/src/main/java/de/morigm/magna/stuff/RunnerStuff.java
@@ -1,0 +1,18 @@
+package de.morigm.magna.stuff;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import de.morigm.magna.api.runner.Runner;
+
+public class RunnerStuff 
+{
+	
+	static
+	{
+		runners = new ArrayList<>();
+	}
+	
+	public static final List<Runner> runners;
+	
+}